Initial version of Quart
Quart is a combination of the Flask API and asyncio. It is based on
the premise that Flask is popular and useful due to its API, yet Flask
itself is unlikely to be directly compatible with asyncio.

This should match the API and functionality of the upcoming Flask 1.0
release. Whilst adding (initially basic) HTTP2 support.