Alter request methods on the Quart app
The order of the methods has been rearranged to match the order in
which the methods are called. The request is also passed into these
methods so as to avoid expensive context stack lookups. This gives
around a 1.03 times speedup. This (request) parameter is kept optional
to keep compatibility with Flask, but is safe to be passed between
Quart methods.