Add ability to use Flask extensions with Quart
This is not 100% guaranteed to work, but relies on Flask and Quart
having the same API, and hence as long as quart.* modules are called
flask.* they should work. It requires an additional import of
quart.flask_ext to add this fake naming.

Flask extensions are one of the major reasons as to why Flask is so
useful, so supported them is a good idea. Whilst it is unlikely this
technical will work perfectly it should cover most use cases.