Rename flask_ext to flask_patch
flask_ext is a legacy Flask system for extensions, and hence using
this name suggests it is the same system rather than a system to
provide Flask imports via a patch of Quart. flask_patch is chosen with
hope that it communicates well that the point is to patch Quart to
provide flask imports. This name however loses the message that it is
worth doing only to support flask extensions.