Alter push promise authority setting
The authority as derived from the hosting setup is almost certianly
wrong in production usages, given I suspect most usages will be behind
some kind of reverse proxy (as is common with Flask).

This change assumes that the request authority that prompted the
response and associated push promises is the correct authority to use.