Implement request time access log atoms and use by default
This allows the request time to be logged. This is defined by the time
the app spends handling the request excluding time spent serving the
request.

This also sets the D atom as present in the default access log, as it
is useful.

The atoms again match the Gunicorn definitions.