Allow for the request body to be streamed
This is the first iteration and defines the API.

This allows users to write route handlers that accept streaming
request bodies. This is very useful when large amounts of data
e.g. files are being transferred. As the data need not be stored in
memory at once.

Note that HTTP/1.1 chunked transfer encoding is supported by Quart.