Add the ability to limit the request body size
This is important as it is a possible method to DOS the server (via
say an out of memory killer). It is implemented such that if you
stream and consume the request body you can never hit the limit.

Note, unlike Flask Quart ships with a 16MB default limit (Flask has no
default limit). I think it is better to have safe defaults, and I
think this is safer.