Bugfix a get_data call
e5a690d420f771492a3b54388caf537a4547b453 introduced an attempt to fix
this, but in fact missed the actual issue which is that if the
Werkzeug HTTPException could not be imported it was set to be
``object``. This meant that all exceptions (as classes) were an
instance of this HTTPException and hence triggered the get_data
code. If the code was correct only the Werkzeug HTTPExceptions would
trigger the get_data call, and would not require an await.