Initial support of host matching and subdomains
Quart, does not need the full host matching and subdomain options
present in Werkzeug as Flask only really allows host matching for the
SERVER_NAME in the app config, rather than for any server name. This
means that a host for the URL Rule is automatically created from the
SERVER_NAME if absent and that the host is only matched against the
request host if host matching is enabled.