Implement a url & host property on requests
This matches the Werkzeug wrapper API, but introduces a complication
of requiring the host header (hence the addition to the testing code).
This header is not required for HTTP < 1.1 so it is set to the app
config SERVER_NAME, if present, otherwise blank. This is the best
knowledge the app can have about the host.