Bugfix h2c upgrade process with request body
Requests with a body and h2c upgrade headers are a pain, as the body
must be fully received over HTTP/1.1 before the upgrade response is
sent, the connection is upgraded and the correct response sent via
HTTP/2. This is a pain as Quart can't both upgrade and wait for the
body.

As h2c is a niche use case (most HTTP/2 clients, including all
browsers don't support it), Quart will take the easy option and ignore
the upgrade header if a request has a body. Should a client really
want this, the client should send a pre-flight options request to
upgrade the connection and then send the actual request with the body.