Add full header control methods to the response wrapper
These match the Werkzeug API and should make Quart more useful and
result in less surprises. Note that the www_authenticate property is
omitted due to its complexity and time constraints. Additionally the
make_conditional functionality will be added with support for 206
responses.

See RFC2616 (Section 14) for reference for the header structures and
parsing.