Bugfix cancel the stream get in Body iteration
Previously when the body is complete the body awaitable would return
and leave the stream.get awaitable unawaited. To add to this the
awaitable was not referenced and lost and would also never
complete. This results in small memory leaks followed by
RuntimeError('Event loop is closed') on closure as the queues attempt
to cancel the awaitables.

This change simply keeps a reference to and cancels the awaitable (as
a future).