Refactor to mitigate DOS attack
This is aimed to mitigate an attack whereby a client requests a route
that is known to stream a large amount of data e.g. a video, and stops
reading from its socket. This results in back-pressure to the Quart
server which would previously just continue to write to the socket and
fill in memory buffers. With enough connections and large enough
streaming responses this could exhaust the host memory and crash the
server.

To mitigate this the Quart Serving code now responds to the asyncio
pause_writing and resume_writing callbacks by clearing and setting an
awaitable event. This event can then be awaited (in a drain call)
during streaming, therefore preventing further writes till the back
pressure is reduced. This is similar to the undocumented
`_drain_helper` method on asyncio protocols.