Add additional websocket route settings
This allows the full suite of host, subdomain and default settings to
be used for websocket routes as well.

This also resolves (for now) the question about OPTIONS and HEAD
requests to websocket routes. Having looked at other implementations I
think the expected response is a 405 method not allowed, which Quart
now does. This is overrideable by defining a normal route with said
methods and the same path.