Change the Flask patch system to ensure the context is preserved
This should ensure that the app, request or websocket context is
available in all flask patched functions. This should therefore fix a
number of the bugs seen when using this system.

This also removes the Flask Patch testing in Quart as it contaminates
the other tests. (Note although pytest-xdist comes close, there seems
to be no nice way to consistently isolate tests).