Initial ASGI support
This adds initial support of ASGI, allowing ASGI servers such as
uvicorn and daphne to be used with Quart. To do so requires the Quart
instance (typically called ``app``) to be wrapped in a ASGIServer as so

    # main.py

    from quart.serving import ASGIServer

    app = Quart(__name__)
    ...
    asgi_app = ASGIServer(app)

and then uvicorn (and others) can be used via

    uvicorn main:asgi_app

Note that this is provisional, and I will likely change (or have to
change) this API.

Uvicorn has a real advantage in that it uses httptools rather than h11
to parse HTTP/1, which is quicker (although less robust). It has the
disadvantage that HTTP/2 is not supported.