Bugfix correctly handle query strings
This follows the ASGI practice of considering the path and
query_string as separate. It ensures that the handling is consistent
and correct. As part of it the fragment attribute is removed from
requests/websockets as it is not useful server side.

This is considered a bug fix even though the Request interface has
changed as I don't see the request interface being used in any
extensions and as it fixes a bug.