Introduce a Headers data structure
This follows the Werkzeug equivalent in allowing bytes to be passed as
keys or values when setting. These are decoded and stored as unicode
alongside the other keys/values. Note not all setting methods (update,
extend) have been included as this is intended to be a niche usage to
match Werkzeug.