Bugfix incorrect endpoint overwrite error for wrapped view functions
A synchronous view function is wrapped with a coroutine in the
add_url_rule method, which results in a different handler for each
addition. This was problematic when the same synchronous view function
is added multiple times for multiple routes as the handler would
differ for each addition and hence result in an overwrite assertion
error.

This commit solves the problem by knowing if a handler is in fact a
wrapper added by Quart and if so comparing against the wrapped view
function.