Bugfix ensure that strict slashes can be disabled
This both fixes the Rule code to ensure strict slashes work correctly,
and the plumbing code to ensure the option can be set. With strict
slashes set to False the request should match if it has a trailing
slash or not, if the route has a trailing slash or not.