Bugfix ensure the ASGI callable returns on completion
Previously the callable would only return on receipt of a disconnect
message, rather than returning either on a disconnect or when the
handling is complete.

Note that there is an intentional optimisation to ignore any
disconnect messages and simply run the handling to completion in the
case of http requests. This is to improve performance, which would
suffer significantly with the multi task approach required for
websockets.