Bugfix ASGI handling to ensure completion
This corrects an assumption in
4d4b1f56692abd51701d6969c2bdd1994ec960e9 which has proven
untrue. Notably that there would not be a legitimate use for an
infinite loop within a route function. This is legitimately required
for SSE and (semi-infinitly) for streaming large files.

This also ensures that the tasks spawned are fully completed or
cancelled before the ASGI callable is exited. This stops Quart from
'leaking' tasks.