Bugfix app.run SSL usage and Hypercorn compatibility
Version 0.4.0 of Hypercorn onwards stopped accepting SSL contexts and
instead accepted the configuration of the context (to build the
context itself). This change was made as the SSL context setup is
quite involved for HTTP/2. In order to fix in Quart the run interface
has been changed to accept the SSL configuration rather than a SSL
context. This also (necessarily) bumps the minimum Hypercorn version
to 0.4.0.