Expand the abort functionality and signature
The Werkzeug version effectively takes the code, and optionally
description and a complete reponse object. As these are the
constructor arguments for the Werkzeug HTTPException class. I've
followed the principle (abort takes HTTPException class constructor
arguments) but this means that the Quart version takes a code and
optionally a description and name.