Bugfix ensure sync functions work with async signals
A sync function when connected to an async signal is wrapped in a
coroutine. Blinker by default takes weak reference to the connected
receiver, which in the case of the coroutine is problematic as it
ceases to exist before sync function does. To work around this the
coroutine is placed onto an attribute of the sync function, thereby
ensure it exists as long as the sync function does. (Ideally I'll
think of a better way).