Rethink the Response Body structure
Before this commit the response body (response attribute on Response
classes) had to be an AsyncIterable that returned bytes. This change
insists that it has to be a ResponseBody subclass. Initially this
just ensures that there is a standard sequence conversion method and
that this pattern can be used,

    async with response.response as body:
        async for data in body:
            ...

which is required in the following commit.