Add the ability to send conditional responses
This allows a response to be made conditional (with a similar API to
Werkzeug) thereby allowing a correct response to range requests. It
differs from Werkzeug in that a maximum partial size can be specified
in order to limit the size of a request (regardless of the requested
range). The primary usage of this is for static files e.g. videos and
hence the ability to seek on a file is utilised for better
performance.