ISSUE-43 multipart form data fails to parse correctly
There are two bugs when the request object tries to parse
a multipart form. First when key is None adding it to a multidict
fails since keys can only be str not None. Second cgi.FieldStorage
needs to have valid value for limit equal to the content-length
to parse the multipart/form-data body correctly in case of files
the file data will be empty if this isn't provided