Refactor push promise API
This adds supports for copying certain request headers to the push
promise (in order to mimic the requester having sent the push
request). The chosen headers match those chosen by mod_h2.

This removes the response push_promises functionality on the basis
that push promises are related to the request, not the response, and
as doing so allows for a better API.

Finally this allows for push promises to be tested using the testing
client (or manually if creating Request classes).