Improve the WebSocket routing
Previously Quart would not understand that both a HTTP and a WebSocket
route with the same path can co-exist (instead chosing the first
defined). This fixes that bug and expands upon it to correct build
URLs for WebSockets (with the correct scheme).

Note that this changes the interface for Map.bind_to_request in a
backwards incompatible way. This should have a limited impact as it
didn't match Werkzeug and it is rarely used in extensions.

Note this deprecates PREFERRED_URL_SCHEME for PREFER_SECURE_URLS
(default False) in the config to ensure the correct scheme is chosen
for WebSocket URLs as well.