Allow returning dictionaries from view functions
A returned dictionary will be jsonified in the app's make_response
function. This allows the following,

    @app.route("/")
    async def api():
        data: Dict[str, Any] = ...
        return data

to return a json response, as equivalent to,

    @app.route("/")
    async def api():
        data: Dict[str, Any] = ...
        return jsonify(data)

The dictionary can be combined with headers (another dictionary) and
or a status code (int) as can be done with str returns.

This follows the latest release of Flask.