Add samesite cookie support
The attribute `samesite` specifies that the browser is not allowed to
send the cookie along with cross-site requests. This helps to mitigate
CSRF attacks. Valid values for this attribute are "Strict" and "Lax".

A new config value, `SESSION_COOKIE_SAMESITE`, that's null by default,
determine the `samesite` attribute for the app's cookies.

Note: this update relies on a feature that is present in Python >= 3.8.

See:

* https://gitlab.com/pgjones/quart/issues/220
* https://bugs.python.org/issue29613