Support ASGI root_paths
This is implemented such that it adds it as a root_path to the paths
that Quart is trying to match against the incoming request. If the
request does not have the root_path at its root then a NotFound
exception is raised on matching. This allows the same app to mounted
multiple times with different root paths as configured by the server.