Run synchronous code via a Thread Pool Executor
This ensures that any synchronous functions passed to Quart are run in
an executor, rather than wrapped in a coroutine and run in the event
loop (as before). This should help users that cannot escape some
synchronous code (presumed to be a third party library) yet still want
to use Quart.

I've chosen to copy the contexts to the exector as discussed in
https://bugs.python.org/issue34014 as I think this is expected and
least surprising.

The ensure_coroutine method is a method so that it can be overriden by
extensions, most notably the Trio extension which cannot use the
asyncio executor system.

The signal version does not provide a simple way to override in
extensions, and it is a bit of a special case. Therefor I've left it
as is - but I intend to remove it in a later version.