Change ensure_coroutine to run_sync
This alters the change made in
fc5a7d862f9b55abc7b1aa9db3380632c9ecf1ce.

Looking into the same change for Quart-Trio introduced run_sync as a
name for this, which I think is much clearer. It makes it clear that
the purpose is to run some synchronous code in a manner that works
with the event loop (hence the fact it is a coroutine function).