Change run_sync to ensure_async
This renames e444d6ce8d9bbfec6f683f96d9e7750420f98cba which itself was
a rename. I think the key is that there is a function to run something
synchronously in a thread ``run_sync`` and a function to ensure that
the callable runs async, ``ensure_async``, (either because it is or
because run_sync has been used).