Adopt black and isort to autoformat the files
This removes any need to figure out how to style the code for
Quart, instead allowing these two commands to do it instead,

    black quart/ tests/ setup.py
    isort --dont-skip __init__.py --apply --recursive quart tests

Note that due to some incompatibility with black and isort some flake8
extensions have been removed and some checks ignored.

This matches the style settings used in my other projected and is
possible now that the 0.6 branch is closed.