Switch from multidict to werkzeug
This is the first stage of a longer term effort to base Quart on
Werkzeug directly. This was avoided in the past due to the heavy WSGI
basis in Werkzeug, lack of type hinting, and concerns about the
maintenance. These are all manageable now, it is much clearer what
Quart can and can't use, there are type stubs, and I help a larger
maintenance team.