Switch to Werkzeug routing
Werkzeug version 1.0 supports websocket rules in the routing, this
allows it to be used in Quart. Doing so means that Quart is closer to
the expected API and functionality of Flask.

Note the Werkzeug exceptions cannot be used as they expected an
environ argument in their get_response method. (No environ in ASGI).