Add signal handling to run method
Hypercorn removed the default signal handling so as to leave it as a
choice for API users. As the run method can make opionated decisions
it makes sense to add signal handling, as it results in a pleasant
shutdown (no errors).

Sadly there is no (obvious) way to add the signal handling and use
asyncio.run, hence the duplication of that code.