Bugfix debug traceback rendering
Following the change to the exception handling in
b8dc496425e90e9cfde236d79eb0bd3293329b2f (to follow the Flask
handling) the traceback rendering was broken. This is because the
traceback was rendered using the app's rendering engine which wasnt
available as the rendering now occurs outside of the app context. This
fixes it by using a specific engine to render traceback templates.