Bugfix ensure any exceptions are raised
Previously if an exception resulted in the handle_request or
handle_message tasks finishing by exception the exception would not be
propagated back to the ASGI server. This fixes it, note that any
outstanding tasks are cancelled first.

Note there is still an issue if both tasks fail (e.g. one whilst
cancelling), but asyncio doesn't have MultiErrors as yet...