Bugfix request host value
Following a change in the ASGI specification the pseudo headers are no
longer present in the scope. Instead, for the host header, the server
adds it from the :authority header. Unless the request is HTTP/1.0 and
there is no host or :authority. In this case Quart adds `""` as the
host header value. This fixes 0c5726ff47cdba03198aceadd174c5bba9b5bacd as
with HTTP/1.0 it would return None rather than `""`.