Bugfix request data handling with Flask-Patch
In Flask the full request body is loaded before any handling is
attempted (ignoring the more recent WSGI stream feature). This
functionality is now replicated in Quart. By doing so this seems to
avoid a nested asyncio bug - whereby the nested loop is stuck in a
loop - which I cannot otherwise understand.