Bugfix prevent local data sharing
Whilst not used this way with Quart, if a Local is created and
assigned a value before being accessed across tasks (requests) it
would erroneously share the values across the tasks. This is because
the dictionary would be mutated, rather than replaced - hence the copy
fix.

See also https://github.com/pallets/werkzeug/pull/1778 for the future
replacement of Quart's locals.