Change AppContext lifespan interaction
Previously an AppContext would be pushed, the startup functions run,
and then popped. This meant that whilst `g` could be used in the
startup functions anything placed on `g` would be missing
thereafter. Instead the AppContext is now pushed during startup and
then popped after shutdown. This should make things easier and clearer
for users.