Major refactor of the testing system
This keeps the vast majority of the functionality whilst ensuring that
middleware is also tested and the testing correctly exercises any
changes to the asgi handling made by the user. It also ensures that
the lifespan is respected and correctly called i.e. the
startup/shutdown testing mimics non-testing.

Note that instead of raising a BadRequest exception testing websockets
against a non-websocket route will instead raise a WebsocketResponse
exception with a 400 (Bad request) response. This is more consistent
with expectations.

Note that to test response streaming actually streams the special
`request` method of the test client should be used, as the other
methods will exhaust the response and concat the bytes as one. This
means that testing infinite streaming routes e.g. SSEs should only be
done via the `request` method.

I'm not convinced that storing the context on the app is the best
thing (when preserving) but at the moment I can't think of a better
solution. The _quart._preserve_context usage mimics Flask, which adds
_flask_.preserve_context to the environ.