Add the ability to close websocket connections
This allows a specific code to be sent to the client informing them
that the connection is now closed.

This is also made possible for testing, however in a backwards
incompatible way - by repurposing the existing close method and adding
a new disconnect method. This also results in the HTTP connection
method name changing - for consistency.