Introduce Protocol typings for customisable classes
This makes it easier for extensions to customise these classes and
satisfy any type checkers. For example Quart-Trio, which must
reimplement many of these classes to avoid any asyncio object
instances (which often create event loops).

Note the protocols are fully defined to help users by documenting the
API. I think this is the better trade off than just defining the parts
of the API required by Quart to function.

Thanks to @McSinyx for the TestAppProtocol code and the idea for this.