Switch to Werkzeug's locals
This is possible now that Werkzeug uses ContextVars - note though the
runtime check incase it doesn't (which shouldn't be possible in
environments where Quart can run).

I've decided to type the globals specifically as the object they proxy
too. I think this is best for users as they expect this API and not
the LocalProxy API. I'm also unsure how to type proxies in Python.