Bugfix unhandled exception handling
Following the change to utilise Werkzeug's exceptions the
handle_exception methods can be updated to have the same logic as
Flask - which ensures the correct exception is received by a catch-all
errorhandler.

Note this also updates the make_response method to accept a
HTTPException. This is a subset of what Flask accepts (any WSGI
callable) but should match the likely use-cases.