Alter ensure_async implementation to support extensions
This allows extensions to override the Quart.ensure_async method and
have the change apply to blueprints as well. Without this change it is
possible for differing blueprints to have differing ensure_async
approaches depending on the extension used - which would likely result
in event-loop blocking issues.

This also allows blueprints to have a custom ensure_async, although
this is a by product rather than an expected use case.

(This follows a very similar commit in Flask)