Bugfix Flask Patch sync wrapping async
Flask extensions will often have sync wrappers that when used with
Quart wrap async functions. This often results in the sync function
returning a coroutine (or an error if the extension cannot be used
with Quart). This coroutine should then be awaited, which is what
this fix ensures in Flask-Patch's ensure_async.