Fix the request blueprint method
Previously for a nested blueprint it would return the parent_name.name
rather than the name of the current blueprint. This is fixed and
another blueprints method is added in order to get the parent_name and
name. The latter is made use of in the app.

I've used a cached_property as this method is called a few times per
request (by the app methods).