Add a lock to the form body parsing
This should fix the situation whereby a user invokes the parsing
simultaneously from multiple tasks, e.g.

    form, files = await asyncio.gather(request.form, request.files)

as only one task can parse at any point. Note that the same could be
argued for the get_data, however that doesn't cache the result and
hence it should be clearer to the user not to consume it from multiple
tasks simultaneously.