Demonstrate how to test streaming routes, refer to relevant sections
With testing tips being scattered across sections, it can sometimes be
really confusing and difficult for the reader to find the relevant
example. Provide a single unified example demonstrating streaming both
the request and the response giving the basic idea, and refer to the
sections talking about this in detail.

Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>