Fix logic for locating and loading `.env`/`.quartenv` files.
The existing logic would (due to logic in python-dotenv) base its
search for a `.env` file in the installation directory of Quart
itself rather than in the working directory of the project the
CLI was being executed in. This meant that the `.env` file wasn't
found and thus wasn't loaded. On the other hand, the path for
the `.quartenv` file was explicitly given as `./.quartenv` and
was thus found.

This change pulls in the logic used in Flask which instead of
relying on python-dotenv's logic directly, or Quart's
semi-hardcoded approach, will use the `find_dotenv` function from
python-dotenv to look for the various files, explicitly telling it
to search in the current working directory.