Switch to github and pallets
This switches the hosting to github and adopts the pallets' setup
excluding the package management and documentation styling. These
exclusions will be discussed by the pallets maintainers about the best
joint approach.

This should help the ultimate aim of merging Flask and Quart in some
way that benefits the users.