Remove usage of LocalStacks
This allows for a more direct and efficient usage of ContextVars,
which in local benchmarking leads to a noticeable performance
improvement.

The _*_ctx_stack globals are directly removed and usage should switch
to the *_context globals and has_*_context functions. If these can't
be used then the _cv_* ContextVar can be used directly.

This requires Werkzeug 2.2.0 of which a pre-release is currently
available.