Adopt the JSON provider interface
This allows the JSON implementation to be fully customised by creating
a new JSONProvider and either changing the Quart.json_provider_class
before initialisation or changing the Quart.json after.

This removes the old json_encoder, json_decoder interfaces and the
config settings with the preference now to use a JSONProvider instead.

This should allow for different json implementations such as orjson to
be used, with a possible performance improvement at the expense of
some functionality.