Don't copy the app context into the background task
This can lead to unexpected errors, for example, if a db connection is
assigned to `g` before the request and closed after it would
potentially persist in the background task after the request and hence
fail. Instead it makes more sense to create an app_context within the
background task's context.