Fix root_path usage
The path should be excluding the root_path, not including it as
before. This matches how WSGI apps work and should match user
expectations.

Note the path is set to a space so as to trigger Quart's routing and a
normal 404 response - this allows users to configure it (rather than
the asgi module returning it's own 404).