Fix flake8 blocking commits
flake8 was blocking any commit during pre-commit check with error :

flake8...................................................................Failed
- hook id: flake8
- exit code: 1

src/quart/app.py:775:13: B028 No explicit stacklevel argument found. The warn method from the warnings module uses a stacklevel of 1 by default. This will only show a stack trace for the line on which the warn method is called. It is therefore recommended to use a stacklevel of 2 or greater to provide more information to the user.